requirements.txt
pandas
numpy
scipy
matplotlib
python-dateutil
requests
selenium
dearpygui